WIP: support H.265 (#33) and MJPEG too
This isn't ready to merge because it's depending on an unreleased
version of `h264-reader` and an unreleased (and WIP) version of retina.
But it appears to work in a quick test.

There's no transcoding, so if the browser/player doesn't support these
formats, they don't play. But it will record and allow downloads, and
it seems to be working with Chrome on macOS at least.